week03-homework-2/pyproject.toml
[project]
name = "week03-homework-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index>=0.14.13",
    "llama-index-core>=0.14.13",
    "llama-index-vector-stores-milvus>=0.9.2",
    "llama-index-llms-openai-like>=0.5.1",
    "llama-index-embeddings-huggingface>=0.3.0",
    "python-dotenv",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "llama-index-graph-stores-neo4j",
]
